move BtnOptions to ui folder
